Refactor + add test
The function handle-keystroke had some logic for the game and some for
the translation of a key to a move. There's now a separate function
make-move that only uses the game.

Also added a test for it and the game-field now depends on the game,
instead of on the board + won. This should make it easier if the game
representation changes.